refacotor: separate interface to other file
